pyproject.toml
[project]
name = "hcaptcha-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "google-generativeai>=0.8.3",
    "pillow>=11.0.0",
    "python-decouple>=3.8",
]
